Split remaining over-cap GUI/CLI files and CSS by concern
Brings the maintainability gate under 300 lines for ResultTable.tsx,
SettingsControls.tsx, SettingsDialog.tsx, ToolForm.tsx,
settings.css/dialogs.css, and their Rust backend counterparts
(files.rs, operations.rs, tui/state/mod.rs), following the split
strategy already established earlier in the v0.3.0 remediation pass.
