+++ apps/netscli-gui/src-tauri/src/commands/files/mod.rs
@@ -1,0 +1,14 @@
+mod artifacts;
+mod export;
+mod pcap_path;
+mod preferences;
+
+pub(crate) use artifacts::{open_saved_artifact, reveal_saved_artifact};
+pub(crate) use export::{
+    ensure_file_size_limit, export_text_file, open_result_bundle, save_result_bundle,
+};
+pub(crate) use pcap_path::resolve_gui_pcap_output_path;
+pub(crate) use preferences::{
+    choose_file_save_default_directory, clear_file_save_default_directory,
+    get_file_save_preferences, set_file_save_ask_each_time,
+};
