+++ apps/netscli-gui/src/components/tools/ToolForm.tsx
@@ -1,17 +1,13 @@
-import { ChevronDown, ChevronUp } from 'lucide-react';
+import { ChevronDown } from 'lucide-react';
 import { useEffect, useRef, useState } from 'react';
 
 import { useRovingFocus } from '../primitives/focus';
 import { useAnchoredPopoverPosition, useOverlayDismiss } from '../primitives/overlay';
 import { TOOL_CONFIG } from '../../tools/registry';
-import type { ToolField, WorkspaceTab } from '../../tools/types';
+import type { WorkspaceTab } from '../../tools/types';
 import type { InterfaceInfo } from '../../types/netscli';
-
-interface SelectOption {
-  value: string;
-  label: string;
-  description?: string;
-}
+import { NumberField } from './NumberField';
+import { fieldsForTab, selectOptionsForField } from './toolFormFields';
 
 interface ToolFormProps {
   tab: WorkspaceTab;
@@ -158,174 +154,3 @@ export function ToolForm({ tab, interfaces = [], onPatchForm, onRun }: ToolFormP
   );
 }
 
-function NumberField({
-  disabled,
-  field,
-  tab,
-  onPatchForm,
-  onRun,
-}: {
-  disabled: boolean;
-  field: ToolField;
-  tab: WorkspaceTab;
-  onPatchForm: (tabId: string, key: string, value: string) => void;
-  onRun: (tabId: string) => void;
-}) {
-  const value = tab.form[field.key] ?? '';
-  const numericValue = numberFromFieldValue(value, field);
-  const min = field.min;
-  const max = field.max;
-  const step = field.step ?? 1;
-
-  function commit(raw: string) {
-    onPatchForm(tab.id, field.key, normalizeNumberFieldValue(raw, field));
-  }
-
-  function stepBy(delta: number) {
-    const next = clampNumber((numericValue ?? numberFromFieldValue(field.placeholder, field) ?? min ?? 0) + delta * step, field);
-    onPatchForm(tab.id, field.key, formatNumberFieldValue(next, field));
-  }
-
-  return (
-    <div className="number-field-control">
-      <input
-        aria-label={field.label}
-        autoCapitalize="off"
-        autoCorrect="off"
-        data-testid={`${tab.kind}-${field.key}-input`}
-        disabled={disabled}
-        inputMode="numeric"
-        max={max}
-        min={min}
-        spellCheck={false}
-        step={step}
-        type="number"
-        value={value}
-        placeholder={field.placeholder}
-        onBlur={(event) => commit(event.currentTarget.value)}
-        onChange={(event) => commit(event.target.value)}
-        onKeyDown={(event) => {
-          if (event.key === 'Enter') {
-            commit(event.currentTarget.value);
-            window.setTimeout(() => onRun(tab.id), 0);
-          }
-        }}
-      />
-      <span className="number-field-steppers">
-        <button
-          aria-label={`Increase ${field.label}`}
-          data-tooltip={`Increase ${field.label}`}
-          disabled={disabled || (max !== undefined && numericValue !== null && numericValue >= max)}
-          tabIndex={-1}
-          type="button"
-          onClick={() => stepBy(1)}
-        >
-          <ChevronUp size={10} />
-        </button>
-        <button
-          aria-label={`Decrease ${field.label}`}
-          data-tooltip={`Decrease ${field.label}`}
-          disabled={disabled || (min !== undefined && numericValue !== null && numericValue <= min)}
-          tabIndex={-1}
-          type="button"
-          onClick={() => stepBy(-1)}
-        >
-          <ChevronDown size={10} />
-        </button>
-      </span>
-    </div>
-  );
-}
-
-function normalizeNumberFieldValue(raw: string, field: ToolField): string {
-  const trimmed = raw.trim();
-  if (!trimmed) return '';
-  const parsed = Number(trimmed);
-  if (!Number.isFinite(parsed)) return '';
-  return formatNumberFieldValue(clampNumber(parsed, field), field);
-}
-
-function numberFromFieldValue(value: string | undefined, field: ToolField): number | null {
-  const normalized = normalizeNumberFieldValue(value ?? '', field);
-  return normalized ? Number(normalized) : null;
-}
-
-function clampNumber(value: number, field: ToolField): number {
-  const step = field.step ?? 1;
-  const stepped = Number.isInteger(step) ? Math.trunc(value) : value;
-  const lowerBounded = field.min === undefined ? stepped : Math.max(field.min, stepped);
-  return field.max === undefined ? lowerBounded : Math.min(field.max, lowerBounded);
-}
-
-function formatNumberFieldValue(value: number, field: ToolField): string {
-  return Number.isInteger(field.step ?? 1) ? String(Math.trunc(value)) : String(value);
-}
-
-function fieldsForTab(tab: WorkspaceTab, fields: ToolField[]): ToolField[] {
-  if (tab.kind !== 'pcap') return fields;
-  const mode = tab.form.mode || 'Capture';
-  if (mode === 'Open File') {
-    return fields.filter((field) => field.key === 'mode' || field.key === 'max_packets');
-  }
-  return fields;
-}
-
-function selectOptionsForField(
-  tab: WorkspaceTab,
-  fieldKey: string,
-  staticOptions: string[] | undefined,
-  interfaces: InterfaceInfo[],
-): SelectOption[] {
-  if (tab.kind === 'pcap' && fieldKey === 'interface') {
-    return pcapInterfaceOptions(interfaces, tab.form[fieldKey]);
-  }
-
-  return (staticOptions ?? []).map((option) => ({ value: option, label: option }));
-}
-
-function pcapInterfaceOptions(interfaces: InterfaceInfo[], currentValue: string | undefined): SelectOption[] {
-  const sorted = [...interfaces].sort((a, b) => interfaceScore(b) - interfaceScore(a));
-  const options = sorted.map((iface) => ({
-    value: iface.name,
-    label: iface.name,
-    description: interfaceDescription(iface),
-  }));
-
-  if (currentValue && !options.some((option) => option.value === currentValue)) {
-    options.unshift({
-      value: currentValue,
-      label: currentValue,
-      description: 'Custom interface',
-    });
-  }
-
-  return options;
-}
-
-function interfaceScore(iface: InterfaceInfo): number {
-  const upScore = iface.is_up ? 100 : 0;
-  const physicalScore = iface.is_loopback || isVirtualInterfaceName(iface.name) ? 0 : 50;
-  const addressScore = iface.ips.length > 0 ? 10 : 0;
-  return upScore + physicalScore + addressScore;
-}
-
-function interfaceDescription(iface: InterfaceInfo): string {
-  const address = preferredInterfaceAddress(iface.ips);
-  const state = iface.is_up ? 'Up' : 'Down';
-  const loopback = iface.is_loopback ? ' - Loopback' : '';
-  return address ? `${address} - ${state}${loopback}` : `${state}${loopback}`;
-}
-
-function preferredInterfaceAddress(ips: string[]): string | null {
-  return (
-    ips.find((ip) => !ip.startsWith('169.254.') && !ip.startsWith('fe80:')) ??
-    ips[0] ??
-    null
-  );
-}
-
-function isVirtualInterfaceName(name: string): boolean {
-  return /tailscale|loopback|bluetooth|vethernet|default switch|wsl|docker|hyper-v|vmware|virtualbox|vmnet|npcap|pseudo-interface/i.test(
-    name,
-  );
-}
