+++ apps/netscli-gui/src/styles/shell/dialogs.css
@@ -1,131 +1,3 @@
-.about-overlay {
-  position: fixed;
-  inset: 0;
-  z-index: var(--z-modal);
-  display: grid;
-  place-items: center;
-  background: rgba(0, 0, 0, 0.35);
-}
-
-.about-dialog {
-  position: relative;
-  width: min(560px, calc(100vw - 32px));
-  display: grid;
-  grid-template-columns: 58px minmax(0, 1fr);
-  gap: 14px 18px;
-  padding: 24px;
-  border: 1px solid var(--border-strong);
-  background: var(--bg-elevated);
-  box-shadow: 0 24px 70px var(--shadow);
-}
-
-.about-close {
-  position: absolute;
-  top: 16px;
-  right: 16px;
-  width: 28px;
-  min-width: 28px;
-  height: 28px;
-  display: grid;
-  place-items: center;
-  border: 1px solid var(--border-subtle);
-  background: var(--bg-input);
-  color: var(--text-secondary);
-  border-radius: 4px;
-  padding: 0;
-  line-height: 0;
-}
-
-.about-close svg {
-  display: block;
-}
-
-.about-close:hover {
-  border-color: var(--border-strong);
-  background: var(--bg-selected);
-  color: var(--text-primary);
-}
-
-.about-mark {
-  width: 58px;
-  height: 58px;
-  display: block;
-  border: 0;
-  background: transparent;
-  box-shadow: none;
-}
-
-.about-heading {
-  align-self: center;
-  min-width: 0;
-  padding-right: 42px;
-}
-
-.about-dialog h2 {
-  margin: 0;
-  color: var(--text-primary);
-  font-size: 16px;
-  font-weight: 600;
-}
-
-.about-version {
-  margin: 2px 0 0;
-  color: var(--text-muted);
-  font-size: 12px;
-}
-
-.about-dialog > p:not(.about-version) {
-  grid-column: 1 / -1;
-  margin: 6px 0 0;
-  color: var(--text-secondary);
-  line-height: 1.55;
-}
-
-.about-meta {
-  grid-column: 1 / -1;
-  display: flex;
-  flex-wrap: wrap;
-  gap: 7px;
-  margin-top: 4px;
-}
-
-.about-meta span {
-  display: inline-flex;
-  align-items: center;
-  gap: 5px;
-  padding: 3px 7px;
-  border: 1px solid var(--border-subtle);
-  color: var(--text-muted);
-  font-size: 11px;
-}
-
-.about-links {
-  grid-column: 1 / -1;
-  display: flex;
-  flex-wrap: wrap;
-  gap: 8px;
-  margin-top: 2px;
-}
-
-.about-links button {
-  height: 30px;
-  display: inline-flex;
-  align-items: center;
-  gap: 7px;
-  border: 1px solid var(--border-subtle);
-  border-radius: 4px;
-  background: var(--bg-input);
-  color: var(--text-primary);
-  padding: 0 10px;
-  cursor: default;
-}
-
-.about-links button:hover {
-  border-color: color-mix(in srgb, var(--mint), transparent 40%);
-  color: var(--mint);
-  background: var(--bg-selected);
-}
-
 .dialog-overlay {
   position: fixed;
   inset: 0;
@@ -217,134 +89,4 @@
 .confirm-dialog-actions button.primary {
   border-color: color-mix(in srgb, var(--mint), transparent 35%);
   color: var(--mint);
-}
-
-.search-overlay {
-  position: fixed;
-  inset: 0;
-  z-index: var(--z-modal);
-  display: grid;
-  justify-items: center;
-  align-items: start;
-  padding: 82px 18px 18px;
-  background: rgba(0, 0, 0, 0.28);
-}
-
-.workspace-search {
-  width: min(680px, calc(100vw - 36px));
-  overflow: hidden;
-  border: 1px solid var(--border-strong);
-  background: var(--bg-elevated);
-  box-shadow: 0 24px 70px var(--shadow);
-}
-
-.workspace-search-input {
-  height: 44px;
-  display: grid;
-  grid-template-columns: auto minmax(0, 1fr) auto;
-  align-items: center;
-  gap: 10px;
-  padding: 0 10px 0 14px;
-  border-bottom: 1px solid var(--border-subtle);
-  color: var(--text-muted);
-}
-
-.workspace-search-input input {
-  width: 100%;
-  height: 100%;
-  border: 0;
-  outline: 0;
-  background: transparent;
-  color: var(--text-primary);
-  font: inherit;
-}
-
-.workspace-search-input input::placeholder {
-  color: var(--text-muted);
-}
-
-.workspace-search-input button {
-  width: 28px;
-  height: 28px;
-  display: grid;
-  place-items: center;
-  border: 1px solid var(--border-subtle);
-  border-radius: 4px;
-  background: var(--bg-input);
-  color: var(--text-secondary);
-  padding: 0;
-}
-
-.workspace-search-input button:hover {
-  border-color: var(--border-strong);
-  background: var(--bg-selected);
-  color: var(--text-primary);
-}
-
-.workspace-search-results {
-  max-height: min(420px, calc(100vh - 180px));
-  overflow: auto;
-  padding: 6px;
-  background:
-    linear-gradient(var(--bg-elevated), var(--bg-elevated)) top / 100% 16px no-repeat local,
-    linear-gradient(var(--bg-elevated), var(--bg-elevated)) bottom / 100% 16px no-repeat local,
-    linear-gradient(to bottom, color-mix(in srgb, var(--shadow), transparent 55%), transparent) top / 100% 12px no-repeat scroll,
-    linear-gradient(to top, color-mix(in srgb, var(--shadow), transparent 55%), transparent) bottom / 100% 12px no-repeat scroll;
-}
-
-.workspace-search-results button {
-  width: 100%;
-  min-height: 46px;
-  display: grid;
-  grid-template-columns: 72px minmax(0, 1fr);
-  align-items: center;
-  gap: 10px;
-  border: 0;
-  background: transparent;
-  color: var(--text-primary);
-  padding: 7px 10px;
-  text-align: left;
-}
-
-.workspace-search-results button:hover,
-.workspace-search-results button.active {
-  background: var(--bg-selected);
-}
-
-.workspace-search-kind {
-  color: var(--text-muted);
-  font-size: 11px;
-  text-transform: uppercase;
-}
-
-.workspace-search-main {
-  min-width: 0;
-  display: grid;
-  gap: 2px;
-}
-
-.workspace-search-main strong,
-.workspace-search-main small {
-  min-width: 0;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-}
-
-.workspace-search-main strong {
-  color: var(--text-primary);
-  font-size: 13px;
-  font-weight: 600;
-}
-
-.workspace-search-main small {
-  color: var(--text-muted);
-  font-size: 11px;
-}
-
-.workspace-search-empty {
-  display: block;
-  padding: 18px;
-  color: var(--text-muted);
-  text-align: center;
 }
