+++ apps/netscli-gui/src/styles/shell/search-dialog.css
@@ -1,0 +1,129 @@
+.search-overlay {
+  position: fixed;
+  inset: 0;
+  z-index: var(--z-modal);
+  display: grid;
+  justify-items: center;
+  align-items: start;
+  padding: 82px 18px 18px;
+  background: rgba(0, 0, 0, 0.28);
+}
+
+.workspace-search {
+  width: min(680px, calc(100vw - 36px));
+  overflow: hidden;
+  border: 1px solid var(--border-strong);
+  background: var(--bg-elevated);
+  box-shadow: 0 24px 70px var(--shadow);
+}
+
+.workspace-search-input {
+  height: 44px;
+  display: grid;
+  grid-template-columns: auto minmax(0, 1fr) auto;
+  align-items: center;
+  gap: 10px;
+  padding: 0 10px 0 14px;
+  border-bottom: 1px solid var(--border-subtle);
+  color: var(--text-muted);
+}
+
+.workspace-search-input input {
+  width: 100%;
+  height: 100%;
+  border: 0;
+  outline: 0;
+  background: transparent;
+  color: var(--text-primary);
+  font: inherit;
+}
+
+.workspace-search-input input::placeholder {
+  color: var(--text-muted);
+}
+
+.workspace-search-input button {
+  width: 28px;
+  height: 28px;
+  display: grid;
+  place-items: center;
+  border: 1px solid var(--border-subtle);
+  border-radius: 4px;
+  background: var(--bg-input);
+  color: var(--text-secondary);
+  padding: 0;
+}
+
+.workspace-search-input button:hover {
+  border-color: var(--border-strong);
+  background: var(--bg-selected);
+  color: var(--text-primary);
+}
+
+.workspace-search-results {
+  max-height: min(420px, calc(100vh - 180px));
+  overflow: auto;
+  padding: 6px;
+  background:
+    linear-gradient(var(--bg-elevated), var(--bg-elevated)) top / 100% 16px no-repeat local,
+    linear-gradient(var(--bg-elevated), var(--bg-elevated)) bottom / 100% 16px no-repeat local,
+    linear-gradient(to bottom, color-mix(in srgb, var(--shadow), transparent 55%), transparent) top / 100% 12px no-repeat scroll,
+    linear-gradient(to top, color-mix(in srgb, var(--shadow), transparent 55%), transparent) bottom / 100% 12px no-repeat scroll;
+}
+
+.workspace-search-results button {
+  width: 100%;
+  min-height: 46px;
+  display: grid;
+  grid-template-columns: 72px minmax(0, 1fr);
+  align-items: center;
+  gap: 10px;
+  border: 0;
+  background: transparent;
+  color: var(--text-primary);
+  padding: 7px 10px;
+  text-align: left;
+}
+
+.workspace-search-results button:hover,
+.workspace-search-results button.active {
+  background: var(--bg-selected);
+}
+
+.workspace-search-kind {
+  color: var(--text-muted);
+  font-size: 11px;
+  text-transform: uppercase;
+}
+
+.workspace-search-main {
+  min-width: 0;
+  display: grid;
+  gap: 2px;
+}
+
+.workspace-search-main strong,
+.workspace-search-main small {
+  min-width: 0;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+}
+
+.workspace-search-main strong {
+  color: var(--text-primary);
+  font-size: 13px;
+  font-weight: 600;
+}
+
+.workspace-search-main small {
+  color: var(--text-muted);
+  font-size: 11px;
+}
+
+.workspace-search-empty {
+  display: block;
+  padding: 18px;
+  color: var(--text-muted);
+  text-align: center;
+}
