+++ apps/netscli-gui/src/tools/presentation/filterFieldLabels.ts
@@ -1,0 +1,66 @@
+export const FIELD_KEY_LABELS: Record<string, string> = {
+  ip: 'IP address',
+  addresses: 'Addresses',
+  avg: 'Avg RTT',
+  hop: 'Hop',
+  record_type: 'Record type',
+  ips: 'Address',
+  loss: 'Packet loss',
+  metric: 'Metric',
+  name: 'Interface',
+  open_ports: 'Open ports',
+  resolver: 'Resolver',
+  service_type: 'Service type',
+  ttl: 'TTL',
+  app: 'App marker',
+  address: 'Address / host',
+  best: 'Best RTT',
+  worst: 'Worst RTT',
+};
+
+export const FIELD_KEY_ALIASES: Record<string, string> = {
+  banner: 'banner',
+  addresses: 'addresses',
+  app: 'app',
+  duration: 'duration',
+  file: 'file',
+  hostname: 'hostname',
+  interface: 'interface',
+  ip: 'ip',
+  ips: 'address',
+  mac: 'mac',
+  metric: 'metric',
+  no: 'no',
+  name: 'interface',
+  open_ports: 'open_ports',
+  packets: 'packets',
+  port: 'port',
+  ports: 'ports',
+  proto: 'proto',
+  protocol: 'protocol',
+  record_type: 'type',
+  service: 'service',
+  source: 'source',
+  state: 'state',
+  status: 'status',
+  destination: 'destination',
+  info: 'info',
+  kind: 'kind',
+  length: 'length',
+  loss: 'loss',
+  time: 'time',
+  hop: 'hop',
+  host: 'host',
+  address: 'address',
+  received: 'received',
+  sent: 'sent',
+  avg: 'avg',
+  best: 'best',
+  worst: 'worst',
+  range: 'range',
+  service_type: 'service_type',
+  ttl: 'ttl',
+  value: 'value',
+  resolver: 'resolver',
+  vendor: 'vendor',
+};
